Simplified speech panel object
