+++ Assets/NameBox.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d88f5e4c4fbaaf3488b9e966d42e658d
+guid: 58def30ed0bd44d719b8c946c46d819e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
