+++ Assets/Prefabs/UI/SpeechPanel.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 37849bb488a16344e8888662757571aa
+guid: 6c2c13b57163244ef922b54f36ad699d
 PrefabImporter:
   externalObjects: {}
   userData:
