+++ Assets/Scripts/Interfaces/IAppearingDialogueController.cs
@@ -1,19 +1,7 @@
 public interface IAppearingDialogueController
 {
-    bool PrintTextInstantly { get; set; }
-
-    void ContinueDialog();
-    void SetTimerValue(WaiterType waiterTypeToChange, float valueToTurnFloat);
-    void ToggleDisableTextSkipping(bool disabled);
-    void ClearAllWaiters();
-    void AutoSkipDialog(bool skip);
-    void HideTextbox();
-}
-
-public interface IAppearingText
-{
-    float DefaultPunctuationDelay { set; }
     float CharactersPerSecond { set; }
+    float DefaultPunctuationDelay { set; }
     bool SkippingDisabled { get; set; }
     bool ContinueDialogue { get; set; }
     bool AutoSkip { get; set; }
