+++ Assets/Scripts/Interfaces/IAppearingDialogueController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b068b27dac04dbd409feee61b91b0e9b
+guid: c20e49fb1787d497eb24306eef714ece
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
