+++ Assets/Scripts/UI/AppearingDialogueController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8b31ee0231f45e143a89b3235cac7439
+guid: 8b20474785ead443faab968137999052
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
