--- Assets/Scripts/UI/AppearingText.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b20474785ead443faab968137999052
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
