--- Assets/Scripts/UI/TextSizeGroup.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 713dfb702e6f76044a4a54a91cb92a53
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
